+++ src/Transformers/LambdaFunction.php
@@ -15,7 +15,7 @@ use Rubix\ML\Datasets\DataFrame;
  * @package     Rubix/ML
  * @author      Andrew DalPino
  */
-class LambdaFunction implements Transformer
+class LambdaFunction implements Transformer, Online
 {
     /**
      * The user specified lambda function.
@@ -40,6 +40,17 @@ class LambdaFunction implements Transformer
      * @return void
      */
     public function fit(DataFrame $dataframe) : void
+    {
+        //
+    }
+
+    /**
+     * Update the fitting of the transformer.
+     *
+     * @param  \Rubix\ML\Datasets\DataFrame  $dataframe
+     * @return void
+     */
+    public function update(DataFrame $dataframe) : void
     {
         //
     }