+++ tests/Transformers/L1NormalizerTest.php
@@ -3,6 +3,7 @@
 namespace Rubix\ML\Tests\Transformers;
 
 use Rubix\ML\Datasets\Unlabeled;
+use Rubix\ML\Transformers\Online;
 use Rubix\ML\Transformers\Transformer;
 use Rubix\ML\Transformers\L1Normalizer;
 use PHPUnit\Framework\TestCase;
@@ -28,6 +29,7 @@ class L1NormalizerTest extends TestCase
     {
         $this->assertInstanceOf(L1Normalizer::class, $this->transformer);
         $this->assertInstanceOf(Transformer::class, $this->transformer);
+        $this->assertInstanceOf(Online::class, $this->transformer);
     }
 
     public function test_transform_fitted()