+++ tests/Transformers/MinMaxNormalizerTest.php
@@ -3,6 +3,7 @@
 namespace Rubix\ML\Tests\Transformers;
 
 use Rubix\ML\Datasets\Unlabeled;
+use Rubix\ML\Transformers\Online;
 use Rubix\ML\Transformers\Transformer;
 use Rubix\ML\Transformers\MinMaxNormalizer;
 use PHPUnit\Framework\TestCase;
@@ -29,6 +30,7 @@ class MinMaxNormalizerTest extends TestCase
     {
         $this->assertInstanceOf(MinMaxNormalizer::class, $this->transformer);
         $this->assertInstanceOf(Transformer::class, $this->transformer);
+        $this->assertInstanceOf(Online::class, $this->transformer);
     }
 
     public function test_transform_fitted()